# Optimize `ToString` implementation for integers