@theemathas 2025-01-30
It seems to me that the tests are testing for a known miscompilation: https://github.com/rust-lang/rust/issues/107975