@theemathas 2025-01-30
I'm guessing that this PR caused the compiler to be able to figure out that printing a number doesn't cause that number to change, allowing more optimizations to happen, causing the miscompilation to behave differently.