@GuillaumeGomez 2025-01-30
The point was testing that calling `to_string` on them somehow changed the value. The new checks ensure that it doesn't and I commented as such. Not too sure what else to do on them.