@GuillaumeGomez 2025-01-30
You were right, just calling another method did the trick. Tests are now back to what they were.