@tgross35 2025-02-10
Oh, there was some confusion on another issue - I'm working on float string conversions, not integers.