@GuillaumeGomez 2025-04-05
Since #138546 was merged, rebased.

r? @Amanieu