@Amanieu 2025-05-13
Considering #138338 is likely to re-use much of this code, I'm much less concerned about the testing.

@bors r+