@rust-log-analyzer 2025-05-13
The job **`x86_64-msvc-1`** failed! Check out the build log: [(web)](https://github.com/rust-lang-ci/rust/actions/runs/15007188611/job/42168632573) [(plain)](https://github.com/rust-lang-ci/rust/commit/4f45c389f0846af9595267f54801a945494d4cc6/checks/42168632573/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain
test [rustdoc] tests\rustdoc\private\private-use.rs ... FAILED
test [rustdoc] tests\rustdoc\private\public-impl-mention-private-generic-46380-2.rs ... FAILED
test [rustdoc] tests\rustdoc\private\traits-in-bodies-private.rs ... FAILED
test [rustdoc] tests\rustdoc\process-termination.rs ... FAILED
##[error]There is not enough space on the disk. : 'C:\actions-runner\cached\_diag\blocks\4f981a31-0bbb-4f3c-8b05-9965fa65a308_39055c4c-82f8-4ffa-8080-4a9fd1e87263.2'
##[error]There is not enough space on the disk. : 'C:\actions-runner\cached\_diag\pages\4f981a31-0bbb-4f3c-8b05-9965fa65a308_7b77d366-83b8-417e-a213-7cf4e4d63b93_1.log'
```

</details>