@marcoieni 2025-05-14
You need to edit this line: https://github.com/rust-lang/rust/blob/414482f6a0d4e7290f614300581a0b55442552a3/src/ci/github-actions/jobs.yml#L479

Now I'm looking at this to try to understand what's going on.

Do you think this PR should increase in some way the required disk size?

I looked at the [latest auto build](https://github.com/rust-lang-ci/rust/actions/runs/14997413854/job/42135067191) and there should be plenty of disk space available:

```
Filesystem            Size  Used Avail Use% Mounted on
C:/Program Files/Git  150G  116G   34G  78% /
D:                    150G   37G  114G  25% /d
```

I wonder if this is a flaky error. Maybe we should just retry before switching to the large runner 🤔

Continuing investigating..