@marcoieni 2025-05-14
There was a new release of the runners where GitHub might have changed something https://github.com/actions/runner/releases/tag/v2.324.0