@GuillaumeGomez 2025-05-14
The size should increase, but not in gigabytes.

Seeing your second message, gonna wait for your confirmation before making a change or retrying.