@marcoieni 2025-05-14
> Successful. Well, retry time I guess?

Since I didn't specify the job list, it tested the wrong job. 

![image](https://github.com/user-attachments/assets/008bc10d-1ac8-4407-8152-e6e21ba160bd)

Anyway, this try was not relevant. As I said, the failure is not related to the content of the PR 👍